requirements.txt
duckduckgo-search
langchain-groq
langchain-google-genai
openai
crewai[all]
crewai_tools
newspaper3k
lxml_html_clean
langchain_community
google-generativeai
pillow
streamlit
transformers
tensorflow
tf-keras
keybert
numpy
groq